do not load ffi provider for jruby
native support can be disabled via the config, native support can just fail to load or
create error later inside the ffi. depending on the environment in which JRuby runs, example
of such failing environment is:
https://github.com/jruby/jruby/issues/1861

that is not really the fault of krypt or FFI but overall it is better to have openssl working on all environments then getting failures here and there.

fixes #50

currently JRuby patched krypt as part default gems like this:
https://github.com/jruby/jruby/blob/jruby-1_7/lib/pom.rb#L212